% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timR3.R
\name{set.timR.3}
\alias{set.timR.3}
\title{Title}
\usage{
set.timR.3(time, unit)
}
\arguments{
\item{time}{numeric to indicate the time}

\item{unit}{character string to indicate the unit: c("secs", "mins")}
}
\description{
Title
}
\examples{
\dontrun{set.timR3(10, "secs")}
}
