% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timR4.R
\name{set.timR.4}
\alias{set.timR.4}
\title{Title}
\usage{
set.timR.4(time, unit)
}
\arguments{
\item{time}{numeric to indicate the time}

\item{unit}{character string to indicate the unit: c("secs", "mins")}
}
\value{

}
\description{
Title
}
